fix: Archive site breaking if site doesn't exist
* ignores the drop-site command if site directory doesn't exist
* ignores rename_site_on_upstream step if upstream file for site doesn't exist

+++ agent/proxy.py
@@ -126,7 +126,8 @@ class Proxy(Server):
     def remove_site_from_upstream(self, upstream, site):
         upstream_directory = os.path.join(self.upstreams_directory, upstream)
         site_file = os.path.join(upstream_directory, site)
-        os.remove(site_file)
+        if os.path.exists(site_file):
+            os.remove(site_file)
 
     @job("Rename Site on Upstream")
     def rename_site_on_upstream_job(
